[VOL-5536] Pull and rebase changes from DT repo

+++ voltha-go-controller/nbi/mvlan.go
@@ -1,22 +1,22 @@
 /* -----------------------------------------------------------------------
  * Copyright 2022-2024 Open Networking Foundation Contributors
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+* http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
  * -----------------------------------------------------------------------
  * SPDX-FileCopyrightText: 2022-2024 Open Networking Foundation Contributors
  * SPDX-License-Identifier: Apache-2.0
  * -----------------------------------------------------------------------
- */
+*/
 
 package nbi
 
@@ -153,7 +153,7 @@ func (iph *MulticastHandle) addMvlan(cntx context.Context, w http.ResponseWriter
 	voltAppIntr = voltApp
 	if err := voltAppIntr.AddMvlanProfile(cntx, config.Name, config.Mvlan, config.PonVlan, config.Groups,
 		config.IsChannelBasedGroup, config.OLTSerialNum,
-		uint32(255), config.Proxy); err != nil {
+		255, config.Proxy); err != nil {
 		logger.Errorw(ctx, "northbound-add-mvlan-failed", log.Fields{"mvlan": config.Name, "Reason": err.Error()})
 		http.Error(w, err.Error(), http.StatusConflict)
 		return
